+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ unreleased
   statements). The previous checks were disabled by default.
 * detect usage of non-integer indices in sys.version_info checks
 * extend Y010 to check async functions in addition to normal functions 
-* introduce Y093 (require using TypeAlias for type aliases)
+* introduce Y026 (require using TypeAlias for type aliases)
 * introduce Y017 (disallows assignments with multiple targets or non-name targets)
 * extend Y001 to cover ParamSpec and TypeVarTuple in addition to TypeVar
 * introduce Y018 (detect unused TypeVars)
@@ -29,6 +29,7 @@ unreleased
 * introduce Y024 (prefer ``typing.NamedTuple`` to ``collections.namedtuple``)
 * introduce Y025 (always alias ``collections.abc.Set``)
 * introduce Y027 (Python 2-incompatible extension of Y022)
+* all errors are now enabled by default
 
 20.10.0
 ~~~~~~~